Added a basic money framework
Players now have a Money value that's determined by several base values. When the Money value is modified, it only modifies the baseMoney value instead of the total. The new total including added Tile values is then recalculated. Other changes include the addition of an Player Id value that's used to corrispond better to other Id's such as GridUnit's ownerId, as well as the removal of turns as paramaters in various funcitons.