Removed the "ResourceType" enum, replacing it with a string implementation
The names of resources were also removed from the CardSubtitle enum and have been replaced by the singular entry "NamedRed". All references to these previous enums have been replaced by references to a new string called "resource" in the card class. This value is error checked against the list in the new ResourceInfo class.  In this class, resource names are kept in a List and their Prices are kept in a Dictionary with their names being the keys. This has not only allowed for easier value lookups from GridUnit objects, but also opens the door for custom resources loaded from a file in the future. One last minor addition is a modifiable "graceRounds" value in GameManager, which is used to determine the number of rounds where no neighbor rules apply.